requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
django-cors-headers==3.12.0
django==4.0.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2; python_version >= '3.5'
